+++ src/main/java/ru/yandex/practicum/filmorate/service/user/UserService.java
@@ -8,17 +8,17 @@ public interface UserService {
 
     User createUser(User user);
 
-    User getUser(int id);
+    User getUser(Long id);
 
     List<User> getUsers();
 
     User updateUser(User user);
 
-    void addFriend(int id, int friendId);
+    void addFriend(Long id, Long friendId);
 
-    void removeFriend(int id, int friendId);
+    void removeFriend(Long id, Long friendId);
 
-    List<User> getFriends(int id);
+    List<User> getFriends(Long id);
 
-    List<User> getCommonFriends(int id, int otherId);
+    List<User> getCommonFriends(Long id, Long otherId);
 }
