+++ src/main/java/ru/yandex/practicum/filmorate/validation/validFilmId/FilmIdValidator.java
@@ -7,13 +7,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
 @Slf4j
-public class FilmIdValidator implements ConstraintValidator<ValidFilmId, Integer> {
+public class FilmIdValidator implements ConstraintValidator<ValidFilmId, Long> {
 
     @Autowired
     private FilmStorage filmStorage;
 
     @Override
-    public boolean isValid(Integer filmId, ConstraintValidatorContext context) {
+    public boolean isValid(Long filmId, ConstraintValidatorContext context) {
         boolean exists = filmStorage.containsId(filmId);
 
         if (!exists) {
